+++ R/frag.layers.cohesion.R
@@ -1,11 +1,11 @@
 .cohesion.for.two.layers <- function(g, layers){
-  v1 <- igraph::V(g)$layers == layers[1]
-  v2 <- igraph::V(g)$layers == layers[2]
+  v1 <- igraph::V(g)$layer == layers[1]
+  v2 <- igraph::V(g)$layer == layers[2]
   
   # extract three subgraphs:
-  g1 <-  igraph::subgraph.edges(g, igraph::E(g)[ igraph::V(g)[v1] %--% igraph::V(g)[v1] ])
-  g2 <-  igraph::subgraph.edges(g, igraph::E(g)[ igraph::V(g)[v2] %--% igraph::V(g)[v2] ])
-  g12 <- igraph::subgraph.edges(g, igraph::E(g)[ igraph::V(g)[v1] %--% igraph::V(g)[v2] ])
+  g1 <-  igraph::subgraph_from_edges(g, igraph::E(g)[ igraph::V(g)[v1] %--% igraph::V(g)[v1] ])
+  g2 <-  igraph::subgraph_from_edges(g, igraph::E(g)[ igraph::V(g)[v2] %--% igraph::V(g)[v2] ])
+  g12 <- igraph::subgraph_from_edges(g, igraph::E(g)[ igraph::V(g)[v1] %--% igraph::V(g)[v2] ])
   
   # compute the cohesion values:
   res1 <-  sum(v1, igraph::E(g1)$weight) / sum(igraph::gorder(g), igraph::E(g)$weight)
@@ -15,7 +15,7 @@
   c(res1, res2)
 }
 
-frag.layers.cohesion <- function(graph, layer.attr){
+frag.layers.cohesion <- function(graph, layer.attr, morphometry=NULL, x=NULL, y=NULL, z=NULL, verbose=TRUE){
   # output : value [0;1].
   # tests:
   .check.frag.graph(graph)
@@ -26,28 +26,35 @@ frag.layers.cohesion <- function(graph, layer.attr){
   
   # extract the user-defined layer attribute and reintegrate it as a vertices attribute named "layer":
   layers <- igraph::vertex_attr(graph, name = layer.attr)
-  igraph::V(graph)$layers <- layers
+  igraph::V(graph)$layer <- layers
   layers <- sort(unique(layers))
-  # Conditionnal tests in function of the number of layers:
-  if(length(layers) < 2){
+  # Conditional tests in function of the number of layers:
+  if(verbose & length(layers) < 2){
     warning("At least two different layers are required.")
     return(c(NA, NA))
   }
   pairs <- utils::combn(layers, 2) 
   
   if(length(layers) == 2){
-    if(is.null(igraph::E(graph)$weight)) stop("The edges must be weighted (using the 'frag.edges.weighting' function).")
+    if(verbose & is.null(igraph::E(graph)$weight)) stop("The edges must be weighted (using the 'frag.edges.weighting' function).")
     results <- .cohesion.for.two.layers(graph, layers)
     results <- matrix(results)
   } else{ # if length(layers) > 2
-    warning("More than 2 layers: the 'frag.edges.weighting' function is applied to each pair of layers.")
-    results <- sapply(1:ncol(pairs), function(x){
-      gsub <- frag.get.layers.pair(graph, layer.attr, c(pairs[1, x], pairs[2, x]))
-      gsub <- frag.edges.weighting(gsub, layer.attr)
-      .cohesion.for.two.layers(gsub, unique(igraph::V(gsub)$layers))
+    message("More than 2 layers: the 'frag.edges.weighting' function is applied to each pair of layers.")
+    results <- sapply(seq_len(ncol(pairs)), function(id){
+      
+      gsub <- frag.get.layers.pair(graph, layer.attr, c(pairs[1, id], pairs[2, id]), verbose = verbose)
+      if(length(unique(igraph::V(gsub)$layer)) == 2){
+        gsub <- frag.edges.weighting(gsub, layer.attr, morphometry, x, y, z, verbose = verbose)
+        # .cohesion.for.two.layers(gsub, unique(igraph::V(gsub)$layers))
+        .cohesion.for.two.layers(gsub, c(pairs[1, id], pairs[2, id]))
+      } else{
+        c(NA, NA)
+      }
     })
   }
   rownames(results) <- c("cohesion1", "cohesion2")
   colnames(results) <- apply(pairs, 2, function(x) paste(x, collapse = "/"))
+  results <- round(results, digits = 4)
   t(results)
 }